+++ src/mock/destination.js
@@ -1,4 +1,4 @@
-import {getRandomValue} from '../utils';
+import {getRandomValue} from '../utils/common';
 import {CITIES, DESCRIPTION} from './const';
 
 function generateDestination () {
